refactor(drive): move index_only_row_commitment into its own module
Re-exported from drive::document so call sites are unchanged; inline
crate:: paths replaced with proper imports in the move.
