+++ packages/rs-drive/src/drive/document/mod.rs
@@ -61,70 +61,19 @@ pub(crate) mod ranked_index_tree_type;
 #[cfg(feature = "server")]
 pub(crate) mod index_level_tree_types;
 
+/// The indexOnly row commitment: the payload every indexOnly terminal item
+/// stores, binding one document's index projections into one logical row
+#[cfg(any(feature = "server", feature = "verify"))]
+pub mod index_only_row_commitment;
+
+#[cfg(feature = "server")]
+pub use index_only_row_commitment::index_only_row_commitment;
+#[cfg(any(feature = "server", feature = "verify"))]
+pub use index_only_row_commitment::INDEX_ONLY_ROW_COMMITMENT_SIZE;
+
 /// How many document history entries to fetch at once. This mirrors contract history
 /// and prevents unbounded history reads.
 pub const MAX_DOCUMENT_HISTORY_FETCH_LIMIT: u16 = 10;
-
-/// The byte length of an indexOnly terminal item's payload: the row
-/// commitment below.
-#[cfg(any(feature = "server", feature = "verify"))]
-pub const INDEX_ONLY_ROW_COMMITMENT_SIZE: u32 = 32;
-
-/// The **row commitment** stored as every indexOnly terminal item's payload:
-/// `hash_double(owner ‖ (name ‖ raw index bytes)* ‖ [$createdAt bytes])`
-/// over ALL of the document's properties in sorted-name order.
-///
-/// This is what binds the independently stored index projections of one
-/// document back into one logical row: a delete recomputes the commitment
-/// from its submitted values and every probed entry must carry it, so a
-/// values tuple spliced from two different creates — even two creates by
-/// the same owner — fails the comparison on whichever entry belongs to the
-/// other row. Without it, entry existence alone cannot distinguish "one
-/// document's projections" from "several documents' projections that
-/// happen to coexist".
-#[cfg(feature = "server")]
-pub fn index_only_row_commitment(
-    document: &Document,
-    document_type: DocumentTypeRef,
-    platform_version: &dpp::version::PlatformVersion,
-) -> Result<[u8; 32], crate::error::Error> {
-    use dpp::data_contract::document_type::DocumentPropertyType;
-    use dpp::document::document_methods::DocumentMethodsV0;
-
-    let owner_id = Some(document.owner_id().to_buffer());
-
-    let mut preimage: Vec<u8> = Vec::with_capacity(128);
-    preimage.extend_from_slice(document.owner_id().as_bytes());
-
-    let mut property_names: Vec<&String> = document_type
-        .flattened_properties()
-        .iter()
-        .filter(|(_, property)| !matches!(property.property_type, DocumentPropertyType::Object(_)))
-        .map(|(name, _)| name)
-        .collect();
-    property_names.sort();
-
-    for property_name in property_names {
-        let raw = document
-            .get_raw_for_document_type(property_name, document_type, owner_id, platform_version)?
-            .ok_or(crate::error::Error::Drive(
-                crate::error::drive::DriveError::CorruptedCodeExecution(
-                    "indexOnly row commitment requires every property value; the parser \
-                     requires them all and the transitions carry them all",
-                ),
-            ))?;
-        preimage.extend_from_slice(property_name.as_bytes());
-        preimage.extend_from_slice(&(raw.len() as u32).to_be_bytes());
-        preimage.extend_from_slice(&raw);
-    }
-
-    if let Some(created_at) = document.created_at() {
-        preimage.extend_from_slice(b"$createdAt");
-        preimage.extend_from_slice(&created_at.to_be_bytes());
-    }
-
-    Ok(dpp::util::hash::hash_double(preimage))
-}
 
 #[cfg(feature = "server")]
 /// Creates a reference to a document.
